refactor: Separate ImageUploadInput into a separate component
